% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/content_areas.R
\name{content_areas}
\alias{content_areas}
\title{Return dataframe}
\usage{
content_areas()
}
\value{
The content areas on the CDC Tracking API.
}
\description{
Print the content areas available on the CDC Tracking Network Data API.
}
\examples{
\dontrun{
content_areas()
}
}
