% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EPHTrackR-package.R
\docType{package}
\name{EPHTrackR-package}
\alias{EPHTrackR}
\alias{EPHTrackR-package}
\title{EPHTrackR: R interface to the CDC National Public Health Tracking Network}
\description{
Provides programmatic access to data hosted on the CDC National Public Health Tracking Network (Tracking Network) via the Tracking Network Data API.
}
\keyword{internal}
