% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geographicLevels.R
\name{geographicLevels}
\alias{geographicLevels}
\title{Find geographic levels}
\usage{
geographicLevels(
  measure = NA,
  indicator = NA,
  content_area = NA,
  format = c("ID", "name", "shortName")
)
}
\arguments{
\item{measure}{specify the measures of interest}

\item{indicator}{specify the indicators of interest}

\item{content_area}{specify the content areas of interest}

\item{format}{indicate whether the measure, indicator and/or content_area variables are ID, name or shortName}
}
\value{
The geographic levels for the specified measures on the CDC Tracking API.
}
\description{
Find geographic levels for specified measures available on the CDC Tracking API.
}
\examples{
\dontrun{
}
}
