% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.R
\name{geography}
\alias{geography}
\title{Find geography}
\usage{
geography(
  measure = NA,
  indicator = NA,
  content_area = NA,
  geo_type = NA,
  geo_type_ID = NA,
  format = c("name", "shortName", "ID"),
  rollup = 0
)
}
\arguments{
\item{measure}{specify the measures of interest}

\item{indicator}{specify the indicators of interest}

\item{content_area}{specify the content areas of interest}

\item{geo_type}{specify the Geographic type.}

\item{geo_type_ID}{specify the Geographic type ID.}

\item{format}{indicate whether the measure, indicator and/or content_area variables are ID, name or shortName}

\item{rollup}{default is 0. Specify whether geographic rollup(1) or not (0).}
}
\value{
The geographies for the specified measures on the CDC Tracking API.
}
\description{
Find geography for specified measures and geographic types available on the CDC Tracking API.
}
\examples{
geo1_id<-geography(measure=370,format="ID")
geo2_id<-geography(measure=c(370,423,707),format="ID")
geo2_name<-geography(measure=c("Number of summertime (May-Sep) heat-related deaths, by year","Number of extreme heat days","Number of months of mild drought or worse per year"),format="name")
geo2_shortName<-geography(measure=c("Number of summertime (May-Sep) heat-related deaths, by year","Number of extreme heat days","Number of months of drought per year"),format="shortName")
geo3_id<-geography(content_area = 25,format="ID")
geo4_shortName<-geography(indicator="Historical Heat Days",content_area ="DR",format="shortName")
geo5_shortName<-geography(indicator="Historical Heat Days",content_area ="DR",geo_type = "County" ,format="shortName")
geo6_shortName<-geography(indicator="Historical Heat Days",content_area ="DR",geo_type_ID = 7,format="shortName")
geo7_name<-geography(measure="Number of summertime (May-Sep) heat-related deaths, by year" ,indicator="Historical Extreme Heat Days and Events",content_area ="Drought",format="name")
}
