% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.R
\name{indicators}
\alias{indicators}
\title{Find Indicators}
\usage{
indicators(content_area = NA, format = c("ID", "name", "shortName"))
}
\arguments{
\item{content_area}{specify the content areas of interest}

\item{format}{indicate whether the content_area variable is an ID, name or shortName}
}
\value{
The indicators in the specified content areas on the CDC Tracking API.
}
\description{
Find the indicators in specified content areas available on the CDC Tracking API.
}
\examples{
\dontrun{
indicators(25,"ID")
indicators("Drought","name")
indicators("DR","shortName")

indicators(c(15,25),"ID")
indicators(c("Climate Change","Drought"),"name")
indicators(c("CC","DR"),"shortName")
all_indicators<-indicators()
}
}
