% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_content_areas.R
\name{list_content_areas}
\alias{list_content_areas}
\title{Return dataframe}
\usage{
list_content_areas()
}
\value{
The content areas on the CDC Tracking API.
}
\description{
Print the content areas available on the CDC Tracking Network Data API.
}
\examples{
\dontrun{
list_content_areas()
}
}
