% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_GeographicTypes.R
\name{list_GeographicTypes}
\alias{list_GeographicTypes}
\title{List available geographic types}
\usage{
list_GeographicTypes(measure, simplified_output = T, token = NULL)
}
\arguments{
\item{measure}{Specifies the measure of interest as an ID or name. IDs should be unquoted; name entries should be quoted strings. It is highly recommended that you entered IDs rather than names to avoid issues with string matching with special characters, spaces, and capitalization etc.}

\item{simplified_output}{If TRUE, a simplified output table is returned. If FALSE, the raw output from the Tracking Network Data API is returned. The default is TRUE.}

\item{token}{An optional argument to submit a Tracking API token acquired from trackingsupport(AT)cdc.gov as a quoted string. It is recommended that you save your token using the tracking_api_token() function so that you don't need to enter your token when you run this function. It will be automatically pulled from you .Renviron file.}
}
\value{
This function returns a list with each element containing a data frame corresponding to a specified measure.
}
\description{
Lists geographic types (e.g., state, county) for specified measure(s). If multiple measures are specified, the results for each are returned as separate list elements.
}
\examples{
\dontrun{
list_GeographicTypes(measure=370)

list_GeographicTypes(measure=c(370,423,707))

list_GeographicTypes(measure=c("Number of summertime (May-Sep) heat-related deaths, by year",
                           "Number of extreme heat days",
                           "Number of months of mild drought or worse per year"))
                 
                 
list_GeographicTypes(measure="Number of summertime (May-Sep) heat-related deaths, by year")
}
}
