% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_geography_types.R
\name{list_geography_types}
\alias{list_geography_types}
\title{Find geographic levels}
\usage{
list_geography_types(
  measure = NA,
  format = c("ID", "name", "shortName"),
  simplified_output = T
)
}
\arguments{
\item{measure}{specify the measures of interest}

\item{format}{indicate whether the listed measure(s) is an ID, name or shortName}

\item{simplified_output}{logical. Determines whether output table is simplified with only relevant columns (TRUE) or the raw output from the Tracking Network API (FALSE)}
}
\value{
The geographic levels for the specified measures on the CDC Tracking API.
}
\description{
Find geographic levels for specified measures available on the Tracking Network API. If multiple measures are specified, the results for each are returned as separate list elements.
}
\examples{
\dontrun{
list_geography_types(measure=370,format="ID")
list_geography_types(measure=c(370,423,707),format="ID")
list_geography_types(measure=c("Number of summertime (May-Sep) heat-related deaths, by year",
                           "Number of extreme heat days",
                           "Number of months of mild drought or worse per year"),
                 format="name")
list_geography_types(measure=c("Number of summertime (May-Sep) heat-related deaths, by year",
                           "Number of extreme heat days","Number of months of drought per year"),
                 format="shortName")
list_geography_types(measure="Number of summertime (May-Sep) heat-related deaths, by year" ,
format="name")
}
}
