% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_stratification_levels.R
\name{list_stratification_levels}
\alias{list_stratification_levels}
\title{Find stratification levels}
\usage{
list_stratification_levels(
  measure = NA,
  geo_type = NA,
  geo_type_ID = NA,
  format = c("name", "shortName", "ID"),
  smoothing = 0
)
}
\arguments{
\item{measure}{specify the measures of interest}

\item{geo_type}{specify the Geographic type.}

\item{geo_type_ID}{specify the Geographic type ID.}

\item{format}{indicate whether the measure, indicator and/or content_area variables are ID, name or shortName}

\item{smoothing}{default is 0. Specify whether data is geographically smoothed(1) or not (0).}
}
\value{
The stratification levels for the specified measures and stratification levels on the CDC Tracking API.
}
\description{
Find stratification levels for specified measures and geographic types available on the  Tracking Network Data API. The output of this function is a list with a separate element for each geography type (e.g. state, county) available for the specified measure. Each row in the data frames contained in the list shows the available stratifications for the measure.
}
\examples{
\dontrun{
list_stratification_levels(measure=370,format="ID")

list_stratification_levels(measure=c(370,423,707),format="ID")

list_stratification_levels(measure=c("Number of summertime (May-Sep) heat-related deaths, by year",
                                "Number of extreme heat days","Number of months of drought per year"),
                      format="shortName")
}
}
