% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_TemporalItems.R
\name{list_TemporalItems}
\alias{list_TemporalItems}
\title{List temporal items}
\usage{
list_TemporalItems(
  measure,
  geo_type = NA,
  format = "ID",
  simplified_output = FALSE,
  token = NULL
)
}
\arguments{
\item{measure}{Specifies the measure of interest as an ID, name, or shortName. IDs should be unquoted; name and shortName entries should be quoted strings. It is recommended that only one measure be submitted.}

\item{geo_type}{An optional argument in which you can specify a geographic type as a quoted string (e.g., "State", "County") or a geographic type ID as an unquoted numeric value (e.g., 1, 2). The "geographicType" and "geographicTypeId" columns in the list_geography_types() output contain a list of potential geo_type entries associated with each measure.}

\item{format}{Indicates whether the measure argument contains entries formatted as an ID, name, or shortName as a quoted string (e.g., "name", "shortName"). The default is ID.}

\item{simplified_output}{If TRUE, a simplified output table is returned. If FALSE, the raw output from the Tracking Network Data API is returned. The default is FALSE.}

\item{token}{An optional argument to submit a Tracking API token acquired from trackingsupport(AT)cdc.gov as a quoted string. It is recommended that you save your token using the tracking_api_token() function so that you don't need to enter your token when you run this function. It will be automatically pulled from you .Renviron file.}
}
\value{
This function returns a list with each element containing a data frame corresponding to each combination of the specified measures and geographic types.
}
\description{
List the temporal items (periods) available for specified measures, geography types, and geographic items.
}
\examples{
\dontrun{


temp <- list_TemporalItems(measure = 99,
                          format="ID")
                          
                          
                          
}
}
