% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\name{measures}
\alias{measures}
\title{Find Measures}
\usage{
measures(
  indicator = NA,
  content_area = NA,
  format = c("ID", "name", "shortName")
)
}
\arguments{
\item{indicator}{specify the indicators of interest.}

\item{content_area}{specify the content areas of interest}

\item{format}{indicate whether the indicator and/or content_area variables are ID, name or shortName}
}
\value{
The measures in the specified indicators and/or content areas on the CDC Tracking API.
}
\description{
Find the measures in specified indicators and/or content areas available on the CDC Tracking API.
}
\examples{
measures(indicator=67,format="ID")
measures(indicator=c(67,79),format = "ID")
measures(indicator=c("Heat-Related Mortality","Historical Extreme Heat Days and Events"),format="name")
measures(indicator=c("Heat Mortality","Historical Heat Days"),format="shortName")
measures(content_area = 25,format="ID")
measures(indicator=67,content_area =25,format="ID")
measures(indicator="Heat-Related Mortality",content_area ="Drought",format="name")
measures(indicator="Heat Mortality",content_area ="DR",format="shortName")
all_measures<-measures()
}
