% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{measures_indicators_CAs}
\alias{measures_indicators_CAs}
\title{Data frame containing all content areas, indicators, and measures}
\format{
A data frame with 674 rows and 9 variables:
\describe{
\item{measure_ID}{Measure unique identifier}
\item{measure_name}{Full measure name}
\item{measure_shortName}{Short measure name}
\item{indicator_ID}{Indicator unique identifier}
\item{indicator_name}{Full indicator name}
\item{indicator_shortName}{Short indicator name}
\item{content_area_ID}{Content area unique identifier}
\item{content_area_name}{Full content area name}
\item{content_area_shortName}{Short content area name}
...
}
}
\source{
\url{https://ephtracking.cdc.gov/}
}
\usage{
measures_indicators_CAs
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
Data frame containing all Tracking Network content areas, indicators, and measures and their associated IDs
}
\keyword{internal}
