% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurestratification.R
\name{measurestratification}
\alias{measurestratification}
\title{Find stratification}
\usage{
measurestratification(
  measure = NA,
  geo_type = NA,
  geo_type_ID = NA,
  format = c("name", "shortName", "ID"),
  smoothing = 0
)
}
\arguments{
\item{measure}{specify the measures of interest}

\item{geo_type}{specify the Geographic type.}

\item{geo_type_ID}{specify the Geographic type ID.}

\item{format}{indicate whether the measure is listed as an ID, name or shortName}

\item{smoothing}{default is 0. Specify whether data is geographically smoothed(1) or not (0).}
}
\value{
The stratification for the specified measures and geographic levels on the CDC Tracking API.
}
\description{
Find stratification for specified measures and geographic types available on the CDC Tracking API.
}
\examples{
\dontrun{
}
}
