% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal.R
\name{temporal}
\alias{temporal}
\title{Find temporal periods available.}
\usage{
temporal(
  measure = NA,
  indicator = NA,
  content_area = NA,
  geo_type = NA,
  geo_type_ID = NA,
  geo_items = NA,
  geo_items_ID = NA,
  format = c("name", "shortName", "ID"),
  geo_filter = 1
)
}
\arguments{
\item{measure}{specify the measures of interest}

\item{indicator}{specify the indicators of interest}

\item{content_area}{specify the content areas of interest}

\item{geo_type}{specify the Geographic type.}

\item{geo_type_ID}{specify the Geographic type ID.}

\item{geo_items}{specify Geographic items by name or abbreviation.}

\item{geo_items_ID}{specify Geographic items by ID.}

\item{format}{indicate whether the measure, indicator and/or content_area variables are ID, name or shortName}

\item{geo_filter}{default is 1. Filter to only retrieve filtered geographic type. Most of the time should equal 1.}
}
\value{
The years for specified measures and geographies on the CDC Tracking API.
}
\description{
Find temporal periods on CDC Tracking API for multiple measures and geographies.
}
\examples{
\dontrun{
temp2_shortName<-temporal(content_area = "DR",geo_items_ID = c(4,32,35),format="shortName")
}
}
