% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracking_api_token.R
\name{tracking_api_token}
\alias{tracking_api_token}
\title{Install a Tracking API token to your .Renviro without the need to reference the token directly in each call.}
\usage{
tracking_api_token(key, overwrite = TRUE, install = TRUE)
}
\arguments{
\item{token}{Specifies the API token that was provided to you by the Tracking Program as a quoted string (e.g., "C6016D98-CGCA-495D-BF0B-EAAE20BADD98"). A key can be acquired by emailing, trackingsupport(AT)cdc.gov. Further information available at https://ephtracking.cdc.gov/apihelp.}
}
\description{
This function adds a Tracking API token to your .Renviron file so it can be called securely without being stored in your code. After you have run this token, the token can be called by running Sys.getenv("TRACKING_API_TOKEN").
}
\references{
This function and accompanying documentation were adapted from the tidycensus package with very little modification.
Walker K, Herman M (2023). tidycensus: Load US Census Boundary and Attribute Data as 'tidyverse' and 'sf'-Ready Data Frames.
R package version 1.3.2, https://walker-data.com/tidycensus/.
}
