% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_inventory.R
\name{update_inventory}
\alias{update_inventory}
\title{Update inventory of content areas, indicators, and measures}
\usage{
update_inventory()
}
\value{
An updated data frame stored in the package's data/ folder and accessible via `data(measures_indicators_CAs)`.
}
\description{
Update the locally stored data frame of content areas, indicators, and measures using the CDC Tracking Newtork Data API.
}
\examples{
\dontrun{
update_inventory()
}
}
