% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_measures.R
\docType{data}
\name{update_measures}
\alias{update_measures}
\alias{CAs_raw}
\title{Update Measures List}
\format{
An object of class \code{response} of length 10.
}
\usage{
CAs_raw
}
\value{
A dataset with the latest updates from CDC Tracking API.
}
\description{
Download the latest list of content areas, indicators, and measures using the CDC Tracking API.
}
\examples{
update_measures()
}
\keyword{datasets}
