% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.R
\name{geography}
\alias{geography}
\title{Find geography}
\usage{
geography(
  measure = NA,
  geo_type = NA,
  geo_type_ID = NA,
  format = c("name", "shortName", "ID"),
  simplified_output = TRUE,
  rollup = 0
)
}
\arguments{
\item{measure}{specify the measures of interest}

\item{geo_type}{specify the Geographic type.}

\item{geo_type_ID}{specify the Geographic type ID.}

\item{format}{indicate whether the measure, indicator and/or content_area variables are ID, name or shortName}

\item{simplified_output}{logical. Determines whether output table is simplified with only relevant columns (TRUE) or the raw output from the Tracking Network API (FALSE)}

\item{rollup}{default is 0. Changing this value to 1 results in returning only parent geographies (e.g. states containing all county-level geographies). It makes no difference if the focal geography is already state. It is unlikely you'll need to change this from the default.}
}
\value{
The geographies for the specified measures on the CDC Tracking API.
}
\description{
Find geographies available for specified measures and geographic types available on the Tracking Network API. If multiple measures or geography types are submitted, the resuls from each one will be returned as a separate list element in the output object.
}
\examples{
\dontrun{
geo1_id<-geography(measure=370,format="ID")

geo2_name<-
  geography(measure=c("Number of summertime (May-Sep) heat-related deaths, by year"),
            format="name")
}
}
