% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography_types.R
\name{geography_types}
\alias{geography_types}
\title{Find geographic levels}
\usage{
geography_types(
  measure = NA,
  indicator = NA,
  content_area = NA,
  format = c("ID", "name", "shortName")
)
}
\arguments{
\item{measure}{specify the measures of interest}

\item{indicator}{specify the indicators of interest}

\item{content_area}{specify the content areas of interest}

\item{format}{indicate whether the measure, indicator and/or content_area variables are ID, name or shortName}
}
\value{
The geographic levels for the specified measures on the CDC Tracking API.
}
\description{
Find geographic levels for specified measures available on the CDC Tracking API.
}
\examples{
\dontrun{
geography_types(measure=370,format="ID")
geography_types(measure=c(370,423,707),format="ID")
geography_types(measure=c("Number of summertime (May-Sep) heat-related deaths, by year",
                           "Number of extreme heat days",
                           "Number of months of mild drought or worse per year"),
                 format="name")
geography_types(measure=c("Number of summertime (May-Sep) heat-related deaths, by year",
                           "Number of extreme heat days","Number of months of drought per year"),
                 format="shortName")
geography_types(content_area = 25,format="ID")
geography_types(indicator="Historical Heat Days",
                 content_area ="DR",format="shortName")
geography_types(measure="Number of summertime (May-Sep) heat-related deaths, by year" ,
                 indicator="Historical Extreme Heat Days and Events",
                 content_area ="Drought",format="name")
}
}
