% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_content_areas.R
\name{list_content_areas}
\alias{list_content_areas}
\title{List all content areas}
\usage{
list_content_areas()
}
\value{
This function returns a data frame containing all content area names, shortNames, and IDs.
}
\description{
Lists the content areas available on the CDC Tracking Network. Content areas are the highest level of categorization on the CDC Tracking Network.
}
\examples{
\dontrun{
list_content_areas()
}
}
