% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_GeographicTypes.R
\name{list_GeographicTypes}
\alias{list_GeographicTypes}
\title{List available geographic types}
\usage{
list_GeographicTypes(
  measure = NA,
  format = "ID",
  simplified_output = T,
  token = NULL
)
}
\arguments{
\item{measure}{Specifies the measure of interest as an ID, name, or shortName. IDs should be unquoted; name and shortName entries should be quoted strings.}

\item{format}{Indicates whether the measure argument contains entries formatted as an ID, name, or shortName as a quoted string (e.g., "name", "shortName"). The default is ID.}

\item{simplified_output}{If TRUE, a simplified output table is returned. If FALSE, the raw output from the Tracking Network Data API is returned. The default is TRUE.}

\item{token}{An optional argument to submit a Tracking API token acquired from trackingsupport(AT)cdc.gov as a quoted string. It is recommended that you save your token using the tracking_api_token() function so that you don't need to enter your token when you run this function. It will be automatically pulled from you .Renviron file.}
}
\value{
This function returns a list with each element containing a data frame corresponding to a specified measure.
}
\description{
Lists geographic types (e.g., state, county) for specified measure(s). If multiple measures are specified, the results for each are returned as separate list elements.
}
\examples{
\dontrun{
list_GeographicTypes(measure=370,format="ID")

list_GeographicTypes(measure=c(370,423,707),format="ID")

list_GeographicTypes(measure=c("Number of summertime (May-Sep) heat-related deaths, by year",
                           "Number of extreme heat days",
                           "Number of months of mild drought or worse per year"),
                 format="name")
                 
list_GeographicTypes(measure=c("Number of summertime (May-Sep) heat-related deaths, by year",
                           "Number of extreme heat days","Number of months of drought per year"),
                 format="shortName")
                 
list_GeographicTypes(measure="Number of summertime (May-Sep) heat-related deaths, by year" ,
format="name")
}
}
