% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_geographies.R
\name{list_geographies}
\alias{list_geographies}
\title{DEPRECATED - List available geographies}
\usage{
list_geographies(
  measure = NA,
  geo_type = NA,
  geo_type_ID = NA,
  format = "ID",
  simplified_output = TRUE,
  rollup = 0
)
}
\arguments{
\item{measure}{Specifies the measure of interest as an ID, name, or shortName. IDs should be unquoted; name and shortName entries should be quoted strings.}

\item{geo_type}{An optional argument in which you can specify a geographic type as a quoted string (e.g., "State", "County"). The "geographicType" column in the list_geography_types() output contains a list of geo_types associated with each measure.}

\item{geo_type_ID}{An optional argument in which you can specify a geographic type ID as an unquoted numeric value (e.g., 1, 2). The "geographicTypeId" column in the list_geography_types() output contains a list of geo_types associated with each measure.}

\item{format}{Indicates whether the measure argument contains entries formatted as an ID, name, or shortName as a quoted string (e.g., "name", "shortName"). The default is ID.}

\item{simplified_output}{If TRUE, a simplified output table is returned. If FALSE, the raw output from the Tracking Network Data API is returned. The default is TRUE.}

\item{rollup}{You probably won't need to change this from the default value of 0. Submitting a value of 1 returns only parent geographies (e.g., states instead of all county-level geographies). This argument does nothing if the focal geography type is already a state.}
}
\value{
This function returns a list with each element containing a data frame corresponding to each combination of the specified measures and geographic types.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Replaced by new more powerful function, list_GeographicItems().
}
\examples{
\dontrun{
geo1_id<-list_geographies(measure=370,format="ID")

geo2_name<-
  list_geographies(measure=c("Number of summertime (May-Sep) heat-related deaths, by year"),
            format="name")
}
}
\keyword{internal}
