% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_indicators.R
\name{list_indicators}
\alias{list_indicators}
\title{List indicators}
\usage{
list_indicators(content_area = NA, format = "ID")
}
\arguments{
\item{content_area}{Optional argument used to specify the content area/s of interest as an ID, name, or shortName. IDs should be unquoted, while name and shortName entries should be quoted strings. Available content areas can be identified using list_content_areas().}

\item{format}{Indicates whether the measure argument contains entries formatted as an ID, name or shortName as a quoted string (e.g., "name", "shortName"). The default is ID.}
}
\value{
This function returns a data frame containing all indicator names, shortNames and IDs contained in the specified content area/s.
}
\description{
List the indicators contained within the specified content area/s. Indicators are groups of highly related measures.
}
\examples{
\dontrun{


list_indicators(25,"ID")

list_indicators("Drought","name")

list_indicators("DR","shortName")

all_indicators<-list_indicators()


}
}
