% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_indicators.R
\name{list_indicators}
\alias{list_indicators}
\title{Find Indicators}
\usage{
list_indicators(content_area = NA, format = c("ID", "name", "shortName"))
}
\arguments{
\item{content_area}{specify the content areas of interest}

\item{format}{indicate whether the content_area variable is an ID, name or shortName}
}
\value{
The indicators in the specified content areas on the CDC Tracking API.
}
\description{
Find the indicators in specified content areas available on the CDC Tracking API.
}
\examples{
\dontrun{
list_indicators(25,"ID")

list_indicators("Drought","name")

list_indicators("DR","shortName")

list_indicators(c(15,25),"ID")

list_indicators(c("Climate Change","Drought"),"name")

list_indicators(c("CC","DR"),"shortName")

all_indicators<-list_indicators()
}
}
