% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_measures.R
\name{list_measures}
\alias{list_measures}
\title{List available measures}
\usage{
list_measures(indicator = NA, content_area = NA, format = "ID")
}
\arguments{
\item{indicator}{Optional argument used to specify the indicator(s) of interest as an ID, name, or shortName. IDs should be unquoted, while name and shortName entries should be quoted strings. Available indicators can be identified using list_indicators().}

\item{content_area}{Optional argument used to specify the content area(s) of interest as an ID, name, or shortName. IDs should be unquoted, while name and shortName entries should be quoted strings. Available content areas can be identified using list_content_areas().}

\item{format}{Indicates whether the measure argument contains entries formatted as an ID, name, or shortName as a quoted string (e.g., "name", "shortName"). The default is ID.}
}
\value{
This function returns a data frame containing the measures within the specified indicator(s) and/or content area(s).
}
\description{
Lists the measures contained within specified indicator(s) and/or content area(s). Measures are the core data product of the Tracking Network.
}
\examples{
\dontrun{


list_measures(indicator=67,format="ID")

list_measures(indicator=c(67,173),format = "ID")

list_measures(indicator="Heat-Related Mortality",
        format="name")

list_measures(content_area = 25,format="ID")

list_measures(indicator=67,content_area =25,format="ID")

list_measures(indicator="Historical Temperature & Heat Index",content_area ="Drought",format="name")

all_measures<-list_measures()


}
}
