% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_temporal.R
\name{list_temporal}
\alias{list_temporal}
\title{DEPRECATED - List temporal periods}
\usage{
list_temporal(
  measure = NA,
  geo_type = NA,
  geo_type_ID = NA,
  geo = NA,
  geo_ID = NA,
  format = "ID",
  simplified_output = TRUE,
  geo_filter = 1
)
}
\arguments{
\item{measure}{Specifies the measure of interest as an ID, name, or shortName. IDs should be unquoted; name and shortName entries should be quoted strings.}

\item{geo_type}{An optional argument in which you can specify a geographic type as a quoted string (e.g., "State", "County"). The "geographicType" column in the list_geography_types() output contains a list of geo_types associated with each measure.}

\item{geo_type_ID}{An optional argument in which you can specify a geographic type ID as an unquoted numeric value (e.g., 1, 2). The "geographicTypeId" column in the list_geography_types() output contains a list of geo_types associated with each measure.}

\item{geo}{An optional argument in which to specify geographies of interest as a quoted string (e.g., "Alabama", "Colorado"). Currently, this argument only accepts states even for county or sub-county geography specifications in the geo_type, geo_type_ID, and strat_level arguments. When a state-level geo argument is submitted with a sub-state geography specification, all sub-state geographies within the state will be returned. The "parentName" column in the list_geography() output contains a list of available geos. If this argument is NULL, all geographies will be included in the output.}

\item{geo_ID}{An optional argument in which to specify the FIPS codes of geographies of interest as unquoted numeric values without leading zeros (e.g., 1, 8). Currently, this argument only accepts states even for county or sub-county geography specifications in the geo_type, geo_type_ID, and strat_level arguments. When a state-level geo argument is submitted with a sub-state geography specification, all sub-state geographies within the state will be returned. The "parentGeographicId" column in the list_geography() output contains a list of geo_IDs. If this argument is NULL, all geographies will be included in the output.}

\item{format}{Indicates whether the measure argument contains entries formatted as an ID, name, or shortName as a quoted string (e.g., "name", "shortName"). The default is ID.}

\item{simplified_output}{If TRUE, a simplified output table is returned. If FALSE, the raw output from the Tracking Network Data API is returned. The default is TRUE.}

\item{geo_filter}{A 1 indicates that the geo/geo_ID arguments contain the parent geography type (e.g., states containing counties of interest). When 0, the child geographies should be specified in the geo/geo_ID arguments. Currently, the default value is 1, and this should not be changed because the child geography types cannot yet be used in the geo/geo_ID arguments.}
}
\value{
This function returns a list with each element containing a data frame corresponding to each combination of the specified measures and geographic types.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Replaced by new more powerful function, list_TemporalItems().
}
\examples{
\dontrun{


temp <- list_temporal(measure = 99,
                          format="ID")
                          
                          
                          
}
}
\keyword{internal}
