% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Measure_Data.R
\name{Measure_Data}
\alias{Measure_Data}
\title{Pull data from API}
\usage{
Measure_Data(
  measure = NA,
  indicator = NA,
  content_area = NA,
  geo_type = NA,
  geo_type_ID = NA,
  geo_items = NA,
  geo_items_ID = NA,
  temporal_ID = NA,
  strat_level = NA,
  strat_level_ID = NA,
  format = c("name", "shortName", "ID"),
  smoothing = 0
)
}
\arguments{
\item{measure}{specify the measures of interest}

\item{indicator}{specify the indicators of interest}

\item{content_area}{specify the content areas of interest}

\item{geo_type}{specify the Geographic type.}

\item{geo_type_ID}{specify the Geographic type ID.}

\item{geo_items}{specify Geographic items by name or abbreviation.}

\item{geo_items_ID}{specify Geographic items by ID.}

\item{strat_level}{specify stratification level by name or abbreviation.}

\item{strat_level_ID}{specify stratification level by ID.}

\item{format}{indicate whether the measure, indicator and/or content_area variables are ID, name or shortName}

\item{smoothing}{default is 0. Specify whether data is geographically smoothed(1) or not (0).}

\item{temporal}{specify the years of interest.}
}
\value{
The specified data from the CDC Tracking API.
}
\description{
Pull data from CDC Tracking API for multiple measures, geographies, stratifications and years.
}
\examples{
\dontrun{
}
}
