% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Measure_Data.R
\name{Measure_Data}
\alias{Measure_Data}
\title{Pull data from API}
\usage{
Measure_Data(
  measure = NA,
  geo_type = NA,
  geo_type_ID = NA,
  geo_items = NA,
  geo_items_ID = NA,
  temporal_period = NA,
  strat_level = NA,
  strat_level_ID = NA,
  format = c("name", "shortName", "ID"),
  smoothing = 0,
  geo_filter = 1,
  simplified_output = T
)
}
\arguments{
\item{measure}{specify the measures of interest}

\item{geo_type}{specify the geography type of the geo_items entry as a quoted string (e.g. "State", "County").}

\item{geo_type_ID}{specify the ID of the geography type of the geo_items entry as a numeric value (e.g. 1, 2).}

\item{geo_items}{specify Geographic items by name or abbreviation.}

\item{geo_items_ID}{specify Geographic items by ID.}

\item{temporal_period}{specify the temporal period(s) of interest with unquoted string.}

\item{strat_level}{specify stratification level by name or abbreviation.}

\item{strat_level_ID}{specify stratification level by ID.}

\item{format}{indicate whether the measure, indicator and/or content_area variables are ID, name or shortName}

\item{smoothing}{default is 0. Request geographically smoothed data. If smoothed data is requested, but is not available the function will produce an error.}

\item{geo_filter}{default is 1. Filter to query based on parent geographic type. This is a crude fix for a problem with the API query and for now don't change.}

\item{simplified_output}{logical. Determines whether output table is simplified with only relevant columns (TRUE) or the full raw output from the Tracking Network API (FALSE)}
}
\value{
The specified data from the CDC Tracking API.
}
\description{
Pull data from CDC Tracking API for multiple measures, geographies, stratifications and years.
}
\examples{
\dontrun{
dat1_id<-
  Measure_Data(measure=370,content_area = 25,
               geo_type_ID = c(1,2),geo_items_ID = c(4,32,35),
               temporal = c(2015,2016),strat_level = c("State","ST_CT"),
               format = "ID")
dat2_shortName<-
  Measure_Data(measure="Number of summertime (May-Sep) heat-related deaths, by year",
               indicator="Historical Drought",geo_type_ID = c(1,2),
               geo_items_ID = c(4,32,35),temporal=2015:2016,
               strat_level_ID = 1:2,format="shortName")
dat3_name<-
  Measure_Data(measure="Number of summertime (May-Sep) heat-related deaths, by year",
               content_area = "Drought",geo_items_ID = c(4,32,35),format="name")
}
}
