% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Measure_Data.R
\name{Measure_Data}
\alias{Measure_Data}
\title{Pull data from API}
\usage{
Measure_Data(
  measure = NA,
  geo_type = NA,
  geo_type_ID = NA,
  geo_items = NA,
  geo_items_ID = NA,
  temporal_period = NA,
  strat_level = NA,
  strat_level_ID = NA,
  format = c("name", "shortName", "ID"),
  smoothing = 0,
  geo_filter = 1,
  simplified_output = T
)
}
\arguments{
\item{measure}{specify the measures of interest. This argument must be included when running the function and it is recommended that only one measure be queried at a time.}

\item{geo_type}{specify the geography type of the geo_items entry as a quoted string (e.g. "State", "County").}

\item{geo_type_ID}{specify the ID of the geography type of the geo_items entry as a numeric value (e.g. 1, 2).}

\item{geo_items}{specify Geographic items by name or abbreviation. At this time, this argument only accepts states even for county or sub-county geograhies. If this argument is not submitted, all geographies will be included in the output.}

\item{geo_items_ID}{specify Geographic items by ID. At this time, this argument only accepts state FIPS codes (without leading zero) even for county or sub-county geograhies. If this argument is not submitted, all geographies will be included in the output.}

\item{temporal_period}{specify the temporal period(s) of interest with unquoted string. If this argument is not submitted, all temporal periods will be included in the output.}

\item{strat_level}{specify stratification level by name or abbreviation. This argument should be included when running the function and it is recommended that only one stratification level be queried at a time.}

\item{strat_level_ID}{specify stratification level by ID. This argument should be included when running the function and it is recommended that only one stratification level be queried at a time.}

\item{format}{indicate whether the measure, indicator and/or content_area variables are ID, name or shortName}

\item{smoothing}{default is 0. Request geographically smoothed data. If smoothed data is requested, but is not available the function will produce an error.}

\item{geo_filter}{default is 1. Filter to query based on parent geographic type. This is a crude fix for a problem with the API query and for now don't change.}

\item{simplified_output}{logical. Determines whether output table is simplified with only relevant columns (TRUE) or the full raw output from the Tracking Network API (FALSE)}
}
\value{
The specified data from the CDC Tracking API.
}
\description{
Pull data from CDC Tracking API for multiple measures, geographies, stratifications and years. It is recommend that you include only one measure and one stratification level in a data query (many geographies and temporal periods may be provided however). The function will likely still work if vectors of multiple measures or stratifications are submitted, but the resulting object will be a nested list and it may be difficult to distingish which list applies to a particular stratification level. The output of function calls with a single measure and statification level is a list with one element containing the relevant data frame.
}
\examples{
\dontrun{
dat1_id<-
  Measure_Data(measure=370,content_area = 25,
               geo_type_ID = c(1,2),geo_items_ID = c(4,32,35),
               temporal = c(2015,2016),strat_level = c("State","ST_CT"),
               format = "ID")
dat2_shortName<-
  Measure_Data(measure="Number of summertime (May-Sep) heat-related deaths, by year",
               indicator="Historical Drought",geo_type_ID = c(1,2),
               geo_items_ID = c(4,32,35),temporal=2015:2016,
               strat_level_ID = 1:2,format="shortName")
dat3_name<-
  Measure_Data(measure="Number of summertime (May-Sep) heat-related deaths, by year",
               content_area = "Drought",geo_items_ID = c(4,32,35),format="name")
}
}
