% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratificationlevel.R
\name{stratificationlevel}
\alias{stratificationlevel}
\title{Find stratification levels}
\usage{
stratificationlevel(
  measure = NA,
  geo_type = NA,
  geo_type_ID = NA,
  format = c("name", "shortName", "ID"),
  smoothing = 0
)
}
\arguments{
\item{measure}{specify the measures of interest}

\item{geo_type}{specify the Geographic type.}

\item{geo_type_ID}{specify the Geographic type ID.}

\item{format}{indicate whether the measure, indicator and/or content_area variables are ID, name or shortName}

\item{smoothing}{default is 0. Specify whether data is geographically smoothed(1) or not (0).}
}
\value{
The stratification levels for the specified measures and stratification levels on the CDC Tracking API.
}
\description{
Find stratification levels for specified measures and geographic types available on the CDC Tracking API.
}
\examples{
\dontrun{
stratificationlevel(measure=370,format="ID")

stratificationlevel(measure=c(370,423,707),format="ID")

stratificationlevel(measure=c("Number of summertime (May-Sep) heat-related deaths, by year",
                                "Number of extreme heat days","Number of months of drought per year"),
                      format="shortName")
}
}
