% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_inventory.R
\name{update_inventory}
\alias{update_inventory}
\title{Update Measures List}
\usage{
update_inventory()
}
\value{
A dataset with the latest updates from CDC Tracking API.
}
\description{
Update the stored list of content areas, indicators, and measures using the CDC Tracking API.
}
\examples{
\dontrun{
update_inventory()
}
}
