% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_inventory.R
\name{update_inventory}
\alias{update_inventory}
\title{Update inventory of content areas, indicators, and measures}
\usage{
update_inventory()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated because automatic inventory updates were incorporated directly within the list_measures() function in version 1.0.0.
}
\examples{
\dontrun{


update_inventory()


}
}
\keyword{internal}
