% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_inventory.R
\name{update_inventory}
\alias{update_inventory}
\title{Update inventory of content areas, indicators, and measures}
\usage{
update_inventory()
}
\value{
An updated data frame stored in the package's data/ folder and accessible via `data(measures_indicators_CAs)`.
}
\description{
This function retrieves the latest inventory of content areas, indicators, and measures from the Tracking Network Data API. This information is used internally by the package to send appropriate API data requests. It is recommended that you run this function prior to each session to ensure that the inventory is up-to-date.
}
\examples{
\dontrun{
update_inventory()
}
}
